Add OpenApi Spec generator for Controller/PageController/RestController Extend Example
