+++ examples/openapi/src/main.rs
@@ -1,6 +1,6 @@
 use rwf::controller::middleware::SecureId;
 use rwf::controller::{Middleware, MiddlewareSet, ModelController, OpenApiController};
-use rwf::macros::generate_openapi_model_controller;
+use rwf::macros::{generate_openapi_model_controller, generate_openapi_specs};
 use rwf::model::migrate;
 use rwf::prelude::*;
 
@@ -35,6 +35,49 @@ impl Default for ProductController {
     }
 }
 
+#[derive(macros::Form, macros::TemplateValue, Clone)]
+struct PageForm {
+    data: String,
+}
+
+#[derive(Default, macros::PageController)]
+struct PageFormController;
+
+/// Try to generate OpenApi Specs for non ModelController
+#[generate_openapi_specs]
+#[async_trait]
+impl PageController for PageFormController {
+    async fn get(&self, request: &Request) -> Result<Response, Error> {
+        render!(request, "templates/page.html")
+    }
+    async fn post(&self, request: &Request) -> Result<Response, Error> {
+        let data: PageForm = request.form()?;
+        let stream = turbo_stream!(request, "templates/page_element.html", "list", "elem" => data)
+            .action("append");
+        Ok(Response::new().turbo_stream(&[stream]))
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, ToSchema, ToResponse)]
+struct Body {
+    data: String,
+    ts: OffsetDateTime,
+}
+#[derive(Default)]
+struct BodyController;
+
+#[generate_openapi_specs(Body)]
+#[async_trait]
+impl Controller for BodyController {
+    async fn handle(&self, request: &Request) -> Result<Response, Error> {
+        let body = Body {
+            data: serde_json::to_string(&request.head())?,
+            ts: OffsetDateTime::now_utc(),
+        };
+        Ok(Response::new().json(body)?)
+    }
+}
+
 #[tokio::main]
 async fn main() -> Result<(), rwf::http::Error> {
     migrate().await?;
@@ -45,10 +88,21 @@ async fn main() -> Result<(), rwf::http::Error> {
     // Create a handler without the overwrite of the id field
     //let apictrl = ProductController {middleware: MiddlewareSet::default()}.crud("/prod");
 
+    // Generate PageController with ApiSpecs for get and post Mehthods
+    let page = route!("/page" => PageFormController);
     // Makes the OpenApi Specs available under /openapi/yaml or /openapi/json
     // Serve an API Browser under /openapi/redoc and /openapi/rapidoc
     let openapi = route!("/openapi" => OpenApiController);
 
-    let routes = vec![apictrl, openapi];
+    // Generate a Controller with ApiSpecs and a hint for the Json Type!
+    let bodyctrl = route!("/body" => BodyController);
+
+    let routes = vec![
+        apictrl,
+        page,
+        openapi,
+        bodyctrl,
+        route!("/turbo_stream" => rwf::controller::TurboStream),
+    ];
     rwf::http::Server::new(routes).launch().await
 }
