+++ examples/openapi/templates/page_element.html
@@ -1,0 +1,1 @@
+<li><%= elem.data %></li>
